requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
APScheduler==3.6.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.1
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
colorama==0.4.6
distlib==0.3.6
elasticsearch==6.8.2
et-xmlfile==1.1.0
exceptiongroup==1.0.0
filelock==3.8.0
flake8==5.0.4
frozenlist==1.3.1
google-api-core==2.10.2
google-api-python-client==2.31.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
greenlet==1.1.3.post0
httplib2==0.20.4
identify==2.5.8
idna==3.4
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.2
nodeenv==1.7.0
openpyxl==3.0.10
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
position==0.1.2
pre-commit==2.20.0
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.42
tdm==0.1.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.64.1
tzdata==2022.5
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.6
yarl==1.8.1
